requirements.txt
asgiref==3.8.1
boto3==1.28.84
botocore==1.31.84
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
cryptography==41.0.5
dacktool==0.0.7
dbstream==0.1.21
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.15
django-allauth==0.61.1
django-cors-headers==4.4.0
django-google-sso==3.3.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
filelock==3.13.1
google-api-core==2.17.1
google-api-python-client==1.7.11
google-auth==2.28.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.0
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.62.0
googleauthentication==0.0.17
grpc-google-iam-v1==0.12.7
grpcio==1.62.0
grpcio-status==1.62.0
httplib2==0.22.0
idna==3.6
jmespath==1.0.1
libcst==1.2.0
loguru==0.7.2
mypy-extensions==1.0.0
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==3.11.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.25.4
virtualenv==20.24.6
